ssys2pov -> graph2pov
